+++ lib/src/oauth2_flows/implicit.dart
@@ -14,6 +14,10 @@ import '../utils.dart';
 // This will be overridden by tests.
 String gapiUrl = 'https://apis.google.com/js/client.js';
 
+// According to the CSP3 spec a nonce must be a valid base64 string.
+// https://w3c.github.io/webappsec-csp/#grammardef-base64-value
+final _noncePattern = new RegExp('^[\\w+\/_-]+[=]{0,2}\$');
+
 /// This class performs the implicit browser-based oauth2 flow.
 ///
 /// It has to be used in two steps:
@@ -67,11 +71,14 @@ class ImplicitFlow {
       }
     };
 
-    var script = new html.ScriptElement();
+    var script = _createScript();
     script.src = '${gapiUrl}?onload=dartGapiLoaded';
     script.onError.first.then((errorEvent) {
       timeout.cancel();
-      completer.completeError(new Exception('Failed to load gapi library.'));
+      if (!completer.isCompleted) {
+        // script loading errors can still happen after timeouts
+        completer.completeError(new Exception('Failed to load gapi library.'));
+      }
     });
     html.document.body.append(script);
 
@@ -80,18 +87,27 @@ class ImplicitFlow {
 
   Future<LoginResult> loginHybrid(
           {bool force: false, bool immediate: false, String loginHint}) =>
-      _login(force, immediate, true, loginHint);
+      _login(force, immediate, true, loginHint, null);
 
   Future<AccessCredentials> login(
-      {bool force: false, bool immediate: false, String loginHint}) async {
-    return (await _login(force, immediate, false, loginHint)).credential;
+      {bool force: false,
+      bool immediate: false,
+      String loginHint,
+      List<ResponseType> responseTypes}) async {
+    return (await _login(force, immediate, false, loginHint, responseTypes))
+        .credential;
   }
 
   // Completes with either credentials or a tuple of credentials and authCode.
   //  hybrid  =>  [AccessCredentials credentials, String authCode]
   // !hybrid  =>  AccessCredentials
-  Future<LoginResult> _login(
-      bool force, bool immediate, bool hybrid, String loginHint) {
+  //
+  // Alternatively, the response types can be set directly if `hybrid` is not
+  // set to `true`.
+  Future<LoginResult> _login(bool force, bool immediate, bool hybrid,
+      String loginHint, List<ResponseType> responseTypes) {
+    assert(hybrid != true || responseTypes?.isNotEmpty != true);
+
     var completer = new Completer<LoginResult>();
 
     var gapi = js.context['gapi']['auth'];
@@ -100,7 +116,11 @@ class ImplicitFlow {
       'client_id': _clientId,
       'immediate': immediate,
       'approval_prompt': force ? 'force' : 'auto',
-      'response_type': hybrid ? 'code token' : 'token',
+      'response_type': responseTypes?.isNotEmpty == true
+          ? responseTypes
+              .map((responseType) => _responseTypeToString(responseType))
+              .join(' ')
+          : hybrid ? 'code token' : 'token',
       'scope': _scopes.join(' '),
       'access_type': hybrid ? 'offline' : 'online',
     };
@@ -117,12 +137,12 @@ class ImplicitFlow {
         var expiresInRaw = jsTokenObject['expires_in'];
         var code = jsTokenObject['code'];
         var error = jsTokenObject['error'];
+        var idToken = jsTokenObject['id_token'];
 
         var expiresIn;
         if (expiresInRaw is String) {
           expiresIn = int.parse(expiresInRaw);
         }
-
         if (error != null) {
           completer.completeError(
               new UserConsentException('Failed to get user consent: $error.'));
@@ -131,10 +151,15 @@ class ImplicitFlow {
             tokenType != 'Bearer') {
           completer.completeError(new Exception(
               'Failed to obtain user consent. Invalid server response.'));
+        } else if (responseTypes?.contains(ResponseType.idToken) == true &&
+            idToken?.isNotEmpty != true) {
+          completer.completeError(
+              new Exception('Expected to get id_token, but did not.'));
         } else {
           var accessToken =
               new AccessToken('Bearer', token, expiryDate(expiresIn));
-          var credentials = new AccessCredentials(accessToken, null, _scopes);
+          var credentials = new AccessCredentials(accessToken, null, _scopes,
+              idToken: idToken);
 
           if (hybrid) {
             if (code == null) {
@@ -158,4 +183,58 @@ class LoginResult {
   final String code;
 
   LoginResult(this.credential, {this.code});
+}
+
+/// Convert [responseType] to string value expected by `gapi.auth.authorize`.
+String _responseTypeToString(ResponseType responseType) {
+  String result;
+
+  switch (responseType) {
+    case ResponseType.code:
+      result = 'code';
+      break;
+
+    case ResponseType.idToken:
+      result = 'id_token';
+      break;
+
+    case ResponseType.permission:
+      result = 'permission';
+      break;
+
+    case ResponseType.token:
+      result = 'token';
+      break;
+
+    default:
+      throw ArgumentError('Unknown response type: $responseType');
+  }
+
+  return result;
+}
+
+/// Creates a script that will run properly when strict CSP is enforced.
+///
+/// More specifically, the script has the correct `nonce` value set.
+final _ScriptFactory _createScript = (() {
+  final nonce = _getNonce();
+  if (nonce == null) return () => new html.ScriptElement();
+
+  return () => new html.ScriptElement()..nonce = nonce;
+})();
+
+typedef html.ScriptElement _ScriptFactory();
+
+/// Returns CSP nonce, if set for any script tag.
+String _getNonce({html.Window window}) {
+  final currentWindow = window ?? html.window;
+  final elements = currentWindow.document.querySelectorAll('script');
+  for (final element in elements) {
+    final nonceValue =
+        (element as html.HtmlElement).nonce ?? element.attributes['nonce'];
+    if (nonceValue != null && _noncePattern.hasMatch(nonceValue)) {
+      return nonceValue;
+    }
+  }
+  return null;
 }