+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.2.7
+
+ - Support for specifying desired `ResponseType`, allowing applications to
+   obtain an `id_token` using `ImplicitBrowserFlow`.
+
+## 0.2.6
+
+- Ignore script loading error after timeout for in-browser implicit login-flow.
+
+## 0.2.5+3
+
+- Support `package:http` `>=0.11.3+17 <0.13.0`.
+
 ## 0.2.5+2
 
 * Support Dart 2.